#6388: parse the /pattern/flags envelope in EO, not Java
compile was a pure Java atom doing three things: validating the
leading/trailing slash, cutting the flag letters off the tail, and
compiling the assembled string. Only the last part genuinely needs
the JVM.

compile is now a real EO formation. It validates the envelope and
builds the (?flags)body search string entirely in EO, at the byte
level (raw!, span!, opens!, last!/scan, body!, flags!, search!) —
not through the character-aware string.slice/length/last-index-of
helpers, which walk UTF-8 recursively and made one compile take
450ms instead of 0.25ms, turning EOscanfTest's per-token loop into
a 41-minute hang. Byte-level cuts are equivalent here since '/' is
ASCII and UTF-8 is self-synchronizing.

The only remaining Java is the new nested atom
EOregex$EOcompile$EOassembled: Pattern.compile(search) and
serialize, routing PatternSyntaxException to the same cant-compile
fallback compile already threads through for the slash-validation
errors. It declares no void of its own, reading search/cant-compile
off compile and pattern off compile's own parent through Phi.RHO —
avoiding the earlier attempt's bug of an unbound, unforwarded void.

First slice of the rewrite this issue asks for; the remaining
@todo chain on EOregex$EOpattern$EOmatch$EOmatched_from_index
(literal matching through the final cutover) is unchanged, since
matching itself still goes through java.util.regex.

Verified: EOregexTest 37/37, EOregexEOcompileTest 2/2,
EOregexEOpatternEOmatchEOmatchedfromindexTest 8/8, EOscanfTest
25/25 (~102s, was a 41-minute hang before the byte-level rewrite),
PhDefaultTest 55/55 (1 pre-existing skip), qulice clean.
