+++ .codacy.yml
@@ -3,4 +3,4 @@
 
 ---
 exclude_paths:
-  - "eo-runtime/src/main/java/org/eolang/EO_string/EOregex$EOcompile.java"
+  - "eo-runtime/src/main/java/org/eolang/EO_string/EOregex$EOcompile$EOassembled.java"
