+++ eo-runtime/src/test/java/org/eolang/PhDefaultTest.java
@@ -8,7 +8,7 @@ import com.yegor256.Together;
 import java.security.SecureRandom;
 import org.cactoos.set.SetOf;
 import org.eolang.EO_org.EO_eolang.EOdummy;
-import org.eolang.EO_string.EOregex$EOcompile;
+import org.eolang.EO_string.EOregex$EOcompile$EOassembled;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
@@ -468,8 +468,8 @@ final class PhDefaultTest {
     void keepsSubPackageInForma() {
         MatcherAssert.assertThat(
             "forma must keep the EO sub-package without its EO marker, but it didnt",
-            new EOregex$EOcompile().forma(),
-            Matchers.equalTo("Φ.string.regex.compile")
+            new EOregex$EOcompile$EOassembled().forma(),
+            Matchers.equalTo("Φ.string.regex.compile.assembled")
         );
     }
